backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-allauth==65.16.1
Pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
